+++ beacon_node/client/Cargo.toml
@@ -14,6 +14,7 @@ eth2 = { workspace = true }
 eth2_config = { workspace = true }
 ethereum_ssz = { workspace = true }
 execution_layer = { workspace = true }
+dummy_el = { path = "../../dummy_el" }
 # TODO(zkproofs): add as a workspace dependency
 zkvm_execution_layer = { path = "../../zkvm_execution_layer" }
 proof_generation_service = { path = "../proof_generation_service" }
