+++ beacon_node/client/src/config.rs
@@ -69,6 +69,7 @@ pub struct Config {
     pub network: network::NetworkConfig,
     pub chain: beacon_chain::ChainConfig,
     pub execution_layer: Option<execution_layer::Config>,
+    pub use_dummy_el: bool,
     pub zkvm_execution_layer: Option<zkvm_execution_layer::ZKVMExecutionLayerConfig>,
     pub trusted_setup: Vec<u8>,
     pub http_api: http_api::Config,
@@ -95,6 +96,7 @@ impl Default for Config {
             network: NetworkConfig::default(),
             chain: <_>::default(),
             execution_layer: None,
+            use_dummy_el: false,
             zkvm_execution_layer: None,
             trusted_setup: get_trusted_setup(),
             beacon_graffiti: GraffitiOrigin::default(),
